refatoração da criação dos pcache, criando somente para os usuários envolvidos na operação

+++ src/protected/application/lib/MapasCulturais/Traits/EntityPermissionCache.php
@@ -45,7 +45,7 @@ trait EntityPermissionCache {
         return $class_name;
     }
     
-    static protected array $createdPermissionCache = [];
+    static protected $createdPermissionCache = [];
 
     function createPermissionsCacheForUsers(array $users = null, $flush = false, $delete_old = true) {
 
@@ -89,11 +89,12 @@ trait EntityPermissionCache {
             }
             
             $app->applyHookBoundTo($this, "{$this->hookPrefix}.permissionCacheUsers", [&$users]);
-
-            if($delete_old && $users){
-                $this->deletePermissionsCache();
-            }
         }
+        
+        if($delete_old && $users){
+            $this->deletePermissionsCache($users);
+        }
+
         $conn = $app->em->getConnection();
         $class_name = $this->getPCacheObjectType();
         $permissions = $this->getPermissionsList();
@@ -159,23 +160,40 @@ trait EntityPermissionCache {
         $this->__enabled = true;
     }
     
-    function deletePermissionsCache(){
+    function deletePermissionsCache($users){
         $app = App::i();
         $conn = $app->em->getConnection();
         $class_name = $this->getPCacheObjectType();
         if(!$this->id){
             return;
         }
-        $conn->executeQuery("DELETE FROM pcache WHERE object_type = '{$class_name}' AND object_id = {$this->id}");
+        
+        $users_ids = implode(',', array_map(function($user) { return $user->id; }, $users));
+
+        $conn->executeQuery("DELETE FROM pcache WHERE object_type = '{$class_name}' AND object_id = {$this->id} AND user_id IN ({$users_ids})");
     }
        
-    function enqueueToPCacheRecreation($skip_extra = false){
+    function enqueueToPCacheRecreation(array $users = []){
         $app = App::i();
-        if($app->isEntityEnqueuedToPCacheRecreation($this) || $this->__skipQueuingPCacheRecreation){
-            return false;
+        if($users) {
+            foreach($users as $user) {
+                if($app->isEntityEnqueuedToPCacheRecreation($this, $user) || $this->__skipQueuingPCacheRecreation){
+                    return false;
+                }
+                
+                $app->enqueueEntityToPCacheRecreation($this, $user);
+            }
+        } else {
+            if($app->isEntityEnqueuedToPCacheRecreation($this) || $this->__skipQueuingPCacheRecreation){
+                return false;
+            }
+            
+            $app->enqueueEntityToPCacheRecreation($this);
         }
-        
-        $app->enqueueEntityToPCacheRecreation($this);
+
+        if($this->usesAgentRelation()){
+            $this->deleteUsersWithControlCache();
+        }
 
         return true;
     }
